third_party/blink/web_tests/platform/generic/svg/custom/focus-event-handling-keyboard-expected.txt

TEST

Test whether focusin and focusout events are dispatched and seen in the focusin/focusout event handlers when using keyboard:

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS focusinSeen is "rect1"
PASS focusinSeen is "g"
PASS focusoutSeen is "rect1"
PASS focusinSeen is "use"
PASS focusoutSeen is "g"
PASS focusinSeen is "usesymbol"
PASS focusoutSeen is "use"
PASS focusinSeen is "switch"
PASS focusoutSeen is "usesymbol"
PASS focusinSeen is "img"
PASS focusoutSeen is "switch"
PASS focusoutSeen is "img"
PASS successfullyParsed is true

TEST COMPLETE
